+++ src/components/config/channels/whatsapp/components/tabs/WebhookTab.vue
@@ -116,9 +116,7 @@
     /* istanbul ignore next */
     mounted() {
       if (this.app.config?.webhook?.method) {
-        this.selectedMethod = [
-          { value: this.app.config?.webhook?.method, label: this.app.config?.webhook?.method },
-        ];
+        this.selectedMethod = this.app.config?.webhook?.method;
       }
       this.mountHeaders();
 
@@ -136,7 +134,7 @@
 
       const urlInput = this.getUrlInputElement();
       if (urlInput !== undefined) {
-        this.validUrl = urlInput?.checkValidity();
+        this.validUrl = this.checkURLValidity(urlInput);
       }
     },
     computed: {
@@ -155,13 +153,7 @@
       /* istanbul ignore next */
       getUrlInputElement() {
         let urlInput;
-        this.$refs.webhookUrl.$children?.forEach((firstLayer) => {
-          firstLayer.$children.forEach((secondLayer) => {
-            if (secondLayer.$el.nodeName === 'INPUT') {
-              urlInput = secondLayer.$el;
-            }
-          });
-        });
+        urlInput = this.$refs.webhookUrl?.modelValue || '';
 
         return urlInput;
       },
@@ -210,7 +202,7 @@
       },
       async saveWebhookInfo() {
         const urlInput = this.getUrlInputElement();
-        if (!urlInput?.checkValidity()) {
+        if (!this.checkURLValidity(urlInput)) {
           this.callModal({
             type: 'error',
             text: this.$t('WhatsApp.config.error.invalid_url'),
@@ -227,7 +219,7 @@
             config: {
               webhook: {
                 url: this.webhookUrl,
-                method: this.selectedMethod,
+                method: this.selectedMethod[0].value,
                 headers,
               },
             },
@@ -249,11 +241,17 @@
         }
 
         this.callModal({
-          type: 'error',
+          type: 'success',
           text: this.$t('WhatsApp.config.success.webhook_update'),
         });
 
         this.$root.$emit('updateGrid');
+      },
+
+      checkURLValidity(value) {
+        const urlRegex = /^(https?:\/\/)?(www\.)?([^\s/$.?#].[^\s]*)$/i;
+
+        return urlRegex.test(value);
       },
       callModal({ text, type }) {
         unnnic.unnnicCallAlert({